fix(workspace): one PRD parser and one bead shape (§13.1)
The artifact viewer carried its own `parsePrdDocument`,
`PRD_TECHNICAL_SECTION_CONFIG`, `ParsedPrdDocument` and `parseBeadsArtifact`
beside the ones in `src/lib/`, and `BeadsApprovalNavigator` reached for the
same bead records through a bare `bead as Record<string, unknown>`. Three
readers of the same two artifacts, none of them agreeing on what a valid one
is.

- `PrdDraftView` now uses `parsePrdDocument` and `PRD_TECHNICAL_SECTION_CONFIG`
  from `@/lib/prdDocument`.
- `ParsedBead` and `parseBeadsArtifact` move into `@/lib/beadsDocument`
  (which held only `BEADS_APPROVAL_FOCUS_EVENT`), and the navigator's outline
  builder uses that type instead of an untyped cast.

**Observable change, and it is not on the plan's signed-off list.** The
viewer's parser accepted anything with an `epics` array; the shared one
requires `artifact: 'prd'` and at least one titled epic. A PRD draft missing
either now falls through to the plain-text fallback instead of rendering as a
structured PRD. §13.1 offered a display-tolerant entry point as the
alternative; the owner chose the single strict contract, so that is what this
does. The consequence, stated plainly: a draft the server rejected as invalid
can no longer be inspected through the structured view — the raw tab and the
text fallback are how you look at one now.

Three tests pin the boundary in `ArtifactContentViewer.test.tsx`: a valid
draft still renders, and drafts missing the `artifact` marker or with no
titled epic no longer do. Both negative tests were verified by mutating the
call site back to the lenient parse — they fail against it and pass against
the shared one.

Impact: `PrdApprovalPane` and the server were already on the strict parser, so
this removes a disagreement rather than creating one. 1,247 client tests pass.
