+++ src/components/navigator/BeadsApprovalNavigator.tsx
@@ -1,7 +1,7 @@
 import { Badge } from '@/components/ui/badge'
 import { useQuery } from '@tanstack/react-query'
 import { QUERY_STALE_TIME_5M } from '@/lib/constants'
-import { BEADS_APPROVAL_FOCUS_EVENT } from '@/lib/beadsDocument'
+import { BEADS_APPROVAL_FOCUS_EVENT, type ParsedBead } from '@/lib/beadsDocument'
 import { apiTicketPath } from '@/lib/apiPaths'
 import { throwIfNotOk } from '@/lib/fetchError'
 import { ApprovalOutlineShell } from './ApprovalOutlineShell'
@@ -20,12 +20,10 @@ interface BeadOutlineItem {
 }
 
 function parseBeadsOutline(data: unknown[]): BeadOutlineItem[] {
-  return data.map((bead, index) => {
-    const record = bead as Record<string, unknown>
-    const id = typeof record.id === 'string' ? record.id : `bead-${index}`
-    const title = typeof record.title === 'string' ? record.title : `Bead ${index + 1}`
-    const deps = record.dependencies as Record<string, unknown> | undefined
-    const blockedBy = Array.isArray(deps?.blocked_by) ? deps.blocked_by.length : 0
+  return (data as ParsedBead[]).map((bead, index) => {
+    const id = typeof bead.id === 'string' ? bead.id : `bead-${index}`
+    const title = typeof bead.title === 'string' ? bead.title : `Bead ${index + 1}`
+    const blockedBy = bead.dependencies?.blocked_by?.length ?? 0
     return { index, id, title, dependencyCount: blockedBy }
   })
 }
